+++ engine/api/golang/kurtosis_engine_api_version/kurtosis_engine_api_version.go
@@ -1,0 +1,6 @@
+package kurtosis_engine_api_version
+
+const (
+	// DO NOT UPDATE THIS VALUE! It will get updated automatically during the release process
+	KurtosisEngineApiVersion = "1.0.0"
+)