+++ engine/docs/changelog.md
@@ -1,5 +1,12 @@
 # TBD
 
+# 1.0.0
+### Changes
+* Refactored the entire structure of this repo to now contain the API
+
+### Breaking Changes
+* The API now lives inside this repo in the `github.com/kurtosis-tech/kurtosis-engine-server/api/golang` package
+
 # 0.6.0
 ### Changes
 * Upgraded to Engine API Lib 0.11.0, which uses Kurt Core API Lib rather than Kurt Client
@@ -88,7 +95,7 @@
 
 # 0.2.0
 ### Features
-* Upgraded to [engine API lib 0.6.0](https://github.com/kurtosis-tech/kurtosis-engine-api-lib/blob/develop/docs/changelog.md#060)
+* Upgraded to engine API lib 0.6.0
 
 ### Breaking Changes
 * `GetEnclave` endpoint is replaced with `GetEnclaves`
@@ -110,7 +117,7 @@
 
 # 0.1.5
 ### Changes
-* Upgrade to [engine API lib 0.4.0](https://github.com/kurtosis-tech/kurtosis-engine-api-lib/blob/develop/docs/changelog.md#040)
+* Upgrade to engine API lib 0.4.0
 
 # 0.1.4
 ### Features